backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
boto3==1.35.0
python-dotenv==1.0.1
pypdf==4.3.0
faiss-cpu==1.8.0
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.9.0
sentence-transformers==3.0.0
torch>=2.0.0
rank-bm25==0.2.2
pyspellchecker==0.8.1
